Another attempt to fix Oauth2 and access tokens.
Problem: Refresh token incorrect after expiry (for picker).
Attempted Fix: Attach user database entry directly to the session, and update it on every databse call.

Actual fix: Im so dumb.
Line 146 on index.js: Was using OG time stamp of when the token was created, rather than the current timestamp to check to see if the token needed to be refreshed (it was never refreshing).

Line 162-163 in index.js: I was mis-using the JSON object, and placing stuff in the wrong area. Now its fixed.

Monitoring fix to see if tokens will finally refresh automatically without the need to sign in again.