Printing Functionality Working
Updated the styling to editor to print the markdown as seen in the editor itself. Using the global "noPrint" class will remove that element, and its children from the print view.

This means, PDF export is now available (if the device can print to PDF of course).

Added a print button on the sidebar for easy access!

ISSUE: Checked Boxes Do Not Appear In Printing View! (at least on chrome)

Also, I have moved development to the development server to make my life a bit easier (no need to keep re-installing node modules etc... on different devices).

Also, I cleaned up some of the packages I wasn't using in node.